- style fixes - readable dates - broke up some huge components into smaller ones
