+++ src/components/GrantsExplorer/index.tsx
@@ -4,7 +4,8 @@ import Countdown from './Countdown'
 import { Glyph, ProgressBar, Button, StickyContent, StickyCanvas, Leaderboard, Spinner, DonationBox } from '@components'
 import { breakpoint, palette, typography } from '@theme'
 import { IGrantsWithProjectAndDonationsFragment } from '@types'
-import { rgba, assetPath } from '@lib'
+import { rgba, dumbDateToString } from '@lib'
+import { Sponsors } from '../HomeParts'
 
 interface IGrantsExplorer {
   grant?: IGrantsWithProjectAndDonationsFragment
@@ -25,7 +26,7 @@ const GrantsExplorer = ({ grant }: IGrantsExplorer) => {
             previous
           </Button>
           <Copy>
-            <Date>{grant.date}</Date>
+            <Date>{dumbDateToString(grant.date)}</Date>
             <Description>Today’s Grant</Description>
           </Copy>
           <Button glyphOnly glyph="arrow" glyphRotation={-90} onClick={() => alert('next')} level={2} disabled>
@@ -42,7 +43,7 @@ const GrantsExplorer = ({ grant }: IGrantsExplorer) => {
               <DataLabel>Raised</DataLabel>
               <AmountRaisedRow>
                 <Glyph glyph="ethereum" level={2} />
-                <AmountRaised>{amountRaised}</AmountRaised>
+                <AmountRaised>{amountRaised.toFixed(3)}</AmountRaised>
                 <Goal>&nbsp;/ {grant.goal} goal</Goal>
               </AmountRaisedRow>
             </div>
@@ -59,10 +60,7 @@ const GrantsExplorer = ({ grant }: IGrantsExplorer) => {
 
           <Leaderboard grantId={grant.id} forceUpdate={updateLeaderBoard} />
 
-          <Sponsors>
-            <Microsoft />
-            <ExtendedReality />
-          </Sponsors>
+          <Sponsors />
         </Body>
       </Wrapper>
     </StyledStickyCanvas>
@@ -161,58 +159,6 @@ const AmountRaisedRow = styled.div`
 
 const AmountRaised = styled.div`
   ${typography.title.l4}
-`
-
-const Sponsors = styled.div`
-  display: flex;
-  flex-direction: column-reverse;
-  justify-content: space-around;
-  align-items: center;
-  gap: 15px;
-  margin-top: 24px;
-  @media only screen and (min-width: ${breakpoint.desktop}px) {
-    flex-direction: row;
-  }
-`
-
-const Microsoft = styled.div`
-  width: 94px;
-  height: 20px;
-  background-image: url(${assetPath('/assets/microsoft.svg')});
-  background-size: contain;
-  background-repeat: no-repeat;
-  @media (prefers-color-scheme: dark) {
-    background-image: url(${assetPath('/assets/microsoft-dark.svg')});
-  }
-  @media only screen and (min-width: ${breakpoint.laptop}px) {
-    width: 150px;
-    height: 24px;
-  }
-  @media only screen and (min-width: ${breakpoint.desktop}px) {
-    width: 150px;
-    height: 65px;
-    background-image: url(${assetPath('/assets/microsoft-presents.svg')});
-    @media (prefers-color-scheme: dark) {
-      background-image: url(${assetPath('/assets/microsoft-presents-dark.svg')});
-    }
-  }
-`
-
-const ExtendedReality = styled.div`
-  width: 200px;
-  height: 50px;
-  @media only screen and (min-width: ${breakpoint.laptop}px) {
-    width: 275px;
-    height: 71px;
-  }
-  @media only screen and (min-width: ${breakpoint.desktop}px) {
-    width: 275px;
-    height: 88px;
-  }
-  background-image: url(${assetPath('/assets/season-1.svg')});
-  @media (prefers-color-scheme: dark) {
-    background-image: url(${assetPath('/assets/season-1-dark.svg')});
-  }
 `
 
 export default GrantsExplorer
