+++ src/components/HomeParts/FeaturedArt/index.tsx
@@ -1,30 +1,18 @@
-import { useEffect, useState, useContext } from 'react'
 import styled from 'styled-components'
-import { BigNumber } from 'ethers'
-import { Icon } from '@components'
-import { palette, breakpoint, typography } from '@theme'
-import { rgba, assert, useReadContract, assetPath, LayoutContext } from '@lib'
-import { raffle } from '@copy/home'
-import { ArtizenArtifactsAbi } from '@contracts'
+import { palette, breakpoint } from '@theme'
+import { rgba } from '@lib'
 import { IGrantsWithProjectAndDonationsFragment } from '@types'
 
 type IFeaturedArt = {
   grant?: IGrantsWithProjectAndDonationsFragment
+  loading: boolean
 }
 
 const FeaturedArt = ({ grant }: IFeaturedArt) => {
-  const { setVisibleModalWithAttrs } = useContext(LayoutContext)
-
-  const getDaysAgoFromDate = (start: number) => {
-    const now = new Date()
-    const oneDay = 1000 * 60 * 60 * 24
-    const diffInTime = now.getTime() - start * 1000
-    const diffInDays = Math.round(diffInTime / oneDay)
-    return diffInDays
-  }
+  // const { setVisibleModalWithAttrs } = useContext(LayoutContext)
 
   if (!grant?.submission?.artifacts) return <></>
-  //TODO: MEssage for EricJ, You got now 3 artifacts, one per edition
+  //TODO: message for EricJ, you got now 3 artifacts, one per edition
   const artworkCommunity = grant?.submission?.artifacts ? grant?.submission?.artifacts[0].artwork : ''
 
   // note: current video NFT ratio is 1:.56
