+++ src/components/HomeParts/FeaturedArtPanel/FeaturedArtPanel.stories.tsx
@@ -1,0 +1,9 @@
+import FeaturedArtPanel from './'
+
+export default {
+  title: 'components/FeaturedArtPanel',
+  component: FeaturedArtPanel,
+  argTypes: {},
+}
+
+export const FeaturedArtPanelComponent = (props: any) => <FeaturedArtPanel {...props} />
