+++ src/components/HomeParts/Sponsors/Sponsors.stories.tsx
@@ -1,0 +1,9 @@
+import Sponsors from './'
+
+export default {
+  title: 'components/Sponsors',
+  component: Sponsors,
+  argTypes: {},
+}
+
+export const SponsorsComponent = (props: any) => <Sponsors {...props} />
