+++ src/components/HomeParts/Sponsors/Sponsors.test.tsx
@@ -1,0 +1,9 @@
+import renderer from 'react-test-renderer'
+import Sponsors from './'
+
+describe('Sponsors', () => {
+  it('renders correctly', () => {
+    const tree = renderer.create(<Sponsors />).toJSON()
+    expect(tree).toMatchSnapshot()
+  })
+})
