+++ src/components/HomeParts/index.ts
@@ -1,3 +1,5 @@
 export { default as ApplyForFundingBlurb } from './ApplyForFundingBlurb'
 export { default as FeaturedArt } from './FeaturedArt'
+export { default as FeaturedArtPanel } from './FeaturedArtPanel'
 export { default as Metrics } from './Metrics'
+export { default as Sponsors } from './Sponsors'
