+++ src/lib/index.ts
@@ -8,6 +8,7 @@ export * from './confirmDonationUrl'
 export * from './cookies'
 export * from './CourierNotification'
 export * from './creditCardHelpers'
+export * from './dateHelpers'
 export * from './LayoutContext'
 export * from './envHelpers'
 export * from './getChainId'
